Better logic for default values.
Behavior for normal inputs is untouched: Is there is a default value, we show it.

But so far we used to ignore saved values for context-based inputs.

So you'd go save TBD in a field. Next time you come edit the form, it wouldnt show up.

Resolving it has a big question mark: If there's a saved value AND there's a valid context-value, which one we show?

Imagine the following scenario:

We go fill a form and save "Year Built" as "TBD".
A few days later you'd fill in "Year Built" in the deal as "1980".
Then you come edit the form again. Should you see TBD or "1980"?

Based on observations with Zareen's demo, it seems that 1980 is the right answer.

Although, if there was no context-value, TBD would still be shown.