#2587 fix(Contacts): fix contact page not loading on Edge
The chain of problem:
contact page -> parse-address -> xregex -> getNativeFlags returns
"undefined" string for regexp flags in Edge and IE, which itself
is because core-js RegExp polyfill has a bug in detecting
RegExp.prototype.flags in polyfilling RegExp.prototype.toString.

This bug in core-js only happens when "@babel/preset-env" is configured to add polyfills based on usage ("useBuiltins" option in is set to "usage"),
which is the case for us.

Related issue: https://github.com/zloirock/core-js/issues/536 (reported
yesterday!)