#2546 fix/refactor(EmailCompose): Show correct messages when email is scheduled
Refactoring:
There where actually 7 places in code for sending emails. Two of them
were handled inside `EmailCompose` component and the rest were handled
in different components, which was resulting in code duplication
in the logic related to handling error and success message and disabling
"Send"/"Save" button in compose drawer.
Now the only place that email is sent (`sendContactsEmail` is called)
is EmailCompose, and those mentioned pieces of logic are boiled down
to one single place.