#2471 fix(Sidenav): fix team switcher sometimes not showing
Cause of the problem:
After login, we fetch teams. Previously the second argument were not
passed which caused `roles` to not be fetched. The reason the bug happens
randomly is because another request for loading teams is triggered
somewhere else and it creates a race condition. Depending on which
requests finish sooner, the bug may or may not happen.

What is changed:
- Now we pass `fetchMembers` to `getUserTeam` action creator when fetching
users after login.
- We no longer mutate store state after fetching teams. Updating state
 is already handled within `getUserTeam`.