fix(TextEditor): create separate plugins for each usage
Previously plugins where created once, outside the scope of TextEditor
component. It creates problems when more than one instance of the
editor is used at the same time, which is the case in contacts page
profile.

draft-js version is also rolled back to 0.10.5, as it seems there are
incompatibilities with rich buttons plugin.