#3325 fix(mc): fix missing url
for #3325. We used to hide panels by overriding the config, which caused an race condition: Trait manager was supposed to append to .gjs-pn-views-conainer, which was added to the panel by some of the commands. By emptying-out the default panels, we prevented those buttons from being added, therefore, we never initiated that container.