chore(build): don't load webpack config on production
webpack config is loaded in development to compile and watch
the app via webpack. In production build we don't
need to load webpack config because we serve an already built
version of the app. However it was always loaded via a top-level
import statement. It's now changed to a `require` call which is
called only in development. This way not only unnecessary code
doesn't load in production process, but also it obviates the need
for adding webpack plugins and other build-time libraries to
`dependencies` (instead of `devDependencies`).