chore: fix build broken problem (#3165)
babel's dynamic-import-node plugin should only be included when running unit tests
We currently use `ci` env for running unit tests but `ci` env is also used when
building the app in CI, and this plugin breaks the build. We should not set env
to `ci` for running unit tests. It's changed to `test` environment.