feat(email-compose): show the result of placeholders when possible (#3700)
when sending a single email, the recipient placeholders are determined
based on the recipients of the email. We show the result of the expressions
as a tooltip in this case for better UX.

Whenever the result of some expressions is changed because of a change in
recipients, we show the tooltip for a moment. Plus, you can always hover
over the expressions to see the result if it's known.