fix(intercom): fix intercom behavior when was not closed before refreshing the page (#4971)
Intercom remembers the open/close state for each user. So when user closes/refreshes the page without closing the intercom (clicking on that X button) intercom would show the chat window next time, and since it's considered hidden by default in our redux the close button wont show and it's impossible to close the widget.